+++ AcquaSenseMobile/components/DailyConsumptionCard.js
