+++ AcquaSenseMobile/package.json
@@ -21,7 +21,9 @@
     "expo-status-bar": "~2.0.1",
     "react": "18.3.1",
     "react-native": "0.76.7",
+    "react-native-chart-kit": "^6.12.0",
     "react-native-linear-gradient": "^2.8.3",
+    "react-native-progress": "^5.0.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-web": "~0.19.13"
